feat(Interaction): add new slider controllable replacing 3d control
The 3D Control Slider has now been deprecated for the new Physics
Slider Controllable. It offers a simpler solution that enables the
script to be enabled and disabled at runtime with correct cleanup.

It no longer auto detects the direction of the slider and works on
the specified operating axis.

Due to the way the Physics Slider script works means the Drawer
3D Control can easily be recreated with the Physics Slider script
and therefore the 3D Control Drawer has also been deprecated.